fix: use original direction when doing dir update
I dreamt this.
So, the keyboard event loop changes the direction,
and it can handle multiple keyboard press events between two update()
calls,
but the "new direction" we are trying to set must be compared
with the original direction.

These two subsequent calls in event_loop should not actually succeed
in updating the direction:
    // Let's say current direction is `Up`
    event_loop {
        // Okay, that's fine
        snake.set_direction(`Left`);
        // But, now this would make snake turn around for 180!
        snake.set_direction(`Down`);
    }

    update_loop { .. }

With this fix, that behaviour is not possible.